MR B8: FileSearch backend extraction
Decompose file_search.py (585 lines) into focused package:
- __init__.py: FileSearchTools facade with glob_files + grep_files
- ripgrep_backend.py: RipgrepBackend subprocess + JSON parsing
- python_backend.py: PythonBackend regex fallback + context formatting
- formatter.py: GrepContextFormatter shared formatting utilities

Backward-compatible: FilesystemTools proxy methods continue to work.
All 2,969 tests pass. Ruff clean.
